Remove unused icons and restructure the content-script
