+++ src/content-script/icons.tsx
