+++ src/content-script/popup-store.ts
