+++ src/content-script/popup.scss
